Reimplement various glu functions
Apple's deprecation of OpenGL creates a different problem here because the 3DView is coded using the old OpenGL way. Once 3DView is adapted to work as the GLView (or merged into it) then the glureimpl.h file won't be required. The license of the original file and thus the code in glureimpl.h is the SGI Free Sofware License B, version 2 and it is compatible with the GPL according to https://www.gnu.org/licenses/license-list.html